bux fixes and observer windows
